Rag_langchain_ui_interface/requirements.txt
Flask
psycopg2
python-dotenv
langchain
langchain-google-genai
langchain-postgres
langchain-groq
